requirements.txt
# Essential dependencies for Vercel deployment
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
SQLAlchemy==2.0.41
python-dotenv==1.1.0
gunicorn==21.2.0
requests==2.32.3
rank-bm25
# Original requirements below
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==2.4.0
attrs==25.3.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.27
botocore==1.38.27
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
eval_type_backport==0.2.0
executing==2.0.0
faiss-cpu==1.11.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
google-ai-generativelanguage
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.5.0
jupyter_core==5.4.0
kiwisolver==1.4.8
Levenshtein==0.27.1
markdownify==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.6
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
nest-asyncio==1.5.8
networkx==3.4.2
numpy==1.26.4
orjson==3.10.18
packaging==23.2
pandas==2.2.3
parso==0.8.3
patchright==1.52.4
pickleshare==0.7.5
pillow==11.2.1
platformdirs==3.11.0
portalocker==2.10.1
prompt-toolkit==3.0.39
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.16.1
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.8.2
python-Levenshtein==0.27.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.1
regex==2024.11.6
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.13.0
screeninfo==0.8.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.12.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
uuid7==0.1.0
voyageai==0.3.2
wcwidth==0.2.8
websockets==10.4
xgboost==3.0.2
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0